Fixes and formatting
 - Rename and remove classes
 - Use LoginRepository interface to test fake repository
