+++ app/src/main/java/com/example/basedaggerapp/domain/LoginUsecase.kt
@@ -3,7 +3,7 @@ package com.example.basedaggerapp.domain
 import com.example.basedaggerapp.LoginRepository
 import javax.inject.Inject
 
-class LoginUsecase @Inject constructor(
+class LoginUseCase @Inject constructor(
   private val repository: LoginRepository
 ) {
   fun login(user: String, password: String): Boolean {
