+++ app/src/main/java/com/example/basedaggerapp/repository/RealRepository.kt
@@ -5,6 +5,11 @@ import javax.inject.Inject
 
 class RealRepository @Inject constructor(): LoginRepository {
   override fun logIn(user: String, password: String): Boolean {
-    return user == "user" && password == "password"
+    return user == USER && password == PASSWORD
+  }
+
+  companion object {
+    const val USER = "user"
+    const val PASSWORD = "password"
   }
 }
