+++ app/src/main/java/com/example/basedaggerapp/ui/main/MainViewModel.kt
@@ -1,9 +1,9 @@
 package com.example.basedaggerapp.ui.main
 
 import androidx.lifecycle.ViewModel
-import com.example.basedaggerapp.domain.LoginUsecase
+import com.example.basedaggerapp.domain.LoginUseCase
 import javax.inject.Inject
 
-class MainViewModel @Inject constructor(private val loginUsecase: LoginUsecase): ViewModel() {
-  fun test(user: String, password: String): Boolean = loginUsecase.login(user, password)
+class MainViewModel @Inject constructor(private val loginUseCase: LoginUseCase): ViewModel() {
+  fun test(user: String, password: String): Boolean = loginUseCase.login(user, password)
 }
