+++ app/src/test/java/com/example/basedaggerapp/di/TestAppComponent.kt
@@ -1,7 +1,8 @@
 package com.example.basedaggerapp.di
 
 import android.content.Context
-import com.example.basedaggerapp.ExampleUnitTest
+import com.example.basedaggerapp.domain.LoginUseCaseTest
+import com.example.basedaggerapp.repository.LoginRepositoryTest
 import dagger.BindsInstance
 import dagger.Component
 import dagger.android.AndroidInjectionModule
@@ -22,5 +23,6 @@ interface TestAppComponent: AppComponent {
   }
 
   fun inject(app: TestMyApp)
-  fun inject(test: ExampleUnitTest)
+  fun inject(test: LoginRepositoryTest)
+  fun inject(test: LoginUseCaseTest)
 }
