+++ app/src/test/java/com/example/basedaggerapp/repository/FakeLoginRepository.kt
@@ -4,7 +4,6 @@ import com.example.basedaggerapp.LoginRepository
 import javax.inject.Inject
 
 class FakeLoginRepository @Inject constructor(): LoginRepository {
-  override fun logIn(user: String, password: String): Boolean {
-    return true
-  }
+  override fun logIn(user: String, password: String): Boolean =
+    user == RealRepository.USER && password == RealRepository.PASSWORD
 }
